Implement word-wrapping in Tidy Console Improved documentation Compiler warnings Tested to build and run on: - Mac OS X - Ubuntu - Windows 10
Regression testing:
- Using the new branch word_wrap, there are zero regressions when tested in
  Mac OS X, Ubuntu, and Windows 10.
- In word_wrap, all of the config files were updated to use --console-width 80,
  so that regression testing won't be dependent on the testers' console width!
- Will update the testing branch after merge.

Finally get rid of the dangerous drop-font-tags.

Removed obsolete TidyHideEndTags, which was only ever a synonym for
TidyOmitOptionalTags.

Give TidyCSSPrefix a proper default.

+++ src/messageobj.c
@@ -384,7 +384,7 @@ TidyMessageArgument TY_(getNextMessageArgument)( TidyMessageImpl message, TidyIt
 
 TidyFormatParameterType TY_(getArgType)( TidyMessageImpl message, TidyMessageArgument* arg )
 {
-    int argNum = (int)*arg;
+    int argNum = (int)(size_t)*arg;
     assert( argNum <= message.argcount );
     
     return message.arguments[argNum].type;
@@ -393,7 +393,7 @@ TidyFormatParameterType TY_(getArgType)( TidyMessageImpl message, TidyMessageArg
 
 ctmbstr TY_(getArgFormat)( TidyMessageImpl message, TidyMessageArgument* arg )
 {
-    int argNum = (int)*arg;
+    int argNum = (int)(size_t)*arg;
     assert( argNum <= message.argcount );
     
     return message.arguments[argNum].format;
@@ -402,7 +402,7 @@ ctmbstr TY_(getArgFormat)( TidyMessageImpl message, TidyMessageArgument* arg )
 
 ctmbstr TY_(getArgValueString)( TidyMessageImpl message, TidyMessageArgument* arg )
 {
-    int argNum = (int)*arg;
+    int argNum = (int)(size_t)*arg;
     assert( argNum <= message.argcount );
     assert( message.arguments[argNum].type == tidyFormatType_STRING);
     
@@ -412,7 +412,7 @@ ctmbstr TY_(getArgValueString)( TidyMessageImpl message, TidyMessageArgument* ar
 
 uint TY_(getArgValueUInt)( TidyMessageImpl message, TidyMessageArgument* arg )
 {
-    int argNum = (int)*arg;
+    int argNum = (int)(size_t)*arg;
     assert( argNum <= message.argcount );
     assert( message.arguments[argNum].type == tidyFormatType_UINT);
 
@@ -422,7 +422,7 @@ uint TY_(getArgValueUInt)( TidyMessageImpl message, TidyMessageArgument* arg )
 
 int TY_(getArgValueInt)( TidyMessageImpl message, TidyMessageArgument* arg )
 {
-    int argNum = (int)*arg;
+    int argNum = (int)(size_t)*arg;
     assert( argNum <= message.argcount );
     assert( message.arguments[argNum].type == tidyFormatType_INT);
 
@@ -432,7 +432,7 @@ int TY_(getArgValueInt)( TidyMessageImpl message, TidyMessageArgument* arg )
 
 double TY_(getArgValueDouble)( TidyMessageImpl message, TidyMessageArgument* arg )
 {
-    int argNum = (int)*arg;
+    int argNum = (int)(size_t)*arg;
     assert( argNum <= message.argcount );
     assert( message.arguments[argNum].type == tidyFormatType_DOUBLE);
